Move methods to UserService and GroupService
- Created a MathUtil class for math method
