+++ skyve-war/src/test/java/modules/admin/Group/GroupServiceH2Test.java
@@ -1,4 +1,4 @@
-package modules.admin;
+package modules.admin.Group;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.CoreMatchers.notNullValue;
@@ -6,14 +6,16 @@ import static org.hamcrest.MatcherAssert.assertThat;
 
 import org.junit.jupiter.api.Test;
 
-import modules.admin.Group.GroupExtension;
+import jakarta.inject.Inject;
 import modules.admin.domain.GroupRole;
 import util.AbstractH2Test;
 
-public class ModulesUtilTest extends AbstractH2Test {
+public class GroupServiceH2Test extends AbstractH2Test {
+	
+	@Inject
+	private transient GroupService groupService;
 
 	@Test
-	@SuppressWarnings("static-method")
 	void configureGroup() {
 		// Create test roles
 		GroupRole testRole1 = GroupRole.newInstance();
@@ -25,16 +27,16 @@ public class ModulesUtilTest extends AbstractH2Test {
 		assertThat(testRole2, is(notNullValue()));
 
 		// New Group
-		GroupExtension newGroup = ModulesUtil.configureGroup("TestGroup", "admin.TestRole1", "admin.TestRole2");
+		GroupExtension newGroup = groupService.configureGroup("TestGroup", "admin.TestRole1", "admin.TestRole2");
 		assertThat(newGroup, is(notNullValue()));
 
 		// Attempting to create new group with name and roles of existing group
-		GroupExtension testExistingGroup = ModulesUtil.configureGroup("TestGroup", "admin.TestRole1", "admin.TestRole2");
+		GroupExtension testExistingGroup = groupService.configureGroup("TestGroup", "admin.TestRole1", "admin.TestRole2");
 		assertThat(testExistingGroup, is(notNullValue()));
 		assertThat(testExistingGroup, is(newGroup));
 
 		// Attempting to create new group with name of existing group but missing role
-		GroupExtension testExistingGroupWithMissingRole = ModulesUtil.configureGroup("TestGroup", "admin.TestRole1");
+		GroupExtension testExistingGroupWithMissingRole = groupService.configureGroup("TestGroup", "admin.TestRole1");
 		assertThat(testExistingGroupWithMissingRole, is(notNullValue()));
 		assertThat(testExistingGroupWithMissingRole, is(testExistingGroup));
 	}
